// src/typings/api.ts
export interface ApiChangeSubscriptionPayload {
    orderType: 'CHANGE_SUBSCRIPTION'
    msisdn: string
    newSubscriptionProductId?: number
    simCardProducts: ApiNewSimCardProduct[]
    addOnProducts: ApiAssignedProductToAdd[]
    simCardProductsToRemove: ApiAssignedProductToRemove[]
    addOnProductsToRemove: ApiAssignedProductToRemove[]
    addOnProductsToChange: ApiAssignedProductToChange[]
}

export interface ApiNewSimCardProduct {
    productId: number
    msisdn: string
    simCard: ApiRegisterSimCard | ApiOrderSimCard | ApiRegisterESimProfile | ApiOrderESimProfile
}

export interface ApiRegisterSimCard {
    actionType: 'REGISTER_EXISTING_SIMCARD'
    iccid: string
}

export interface ApiOrderSimCard {
    actionType: 'ORDER_NEW_SIMCARD'
    simCardTypeId: number
    handlingInstructionId: number
    deliveryAddress: Address
}

export interface ApiRegisterESimProfile {
    actionType: 'REGISTER_EXISTING_ESIM_PROFILE'
    iccid: string
    eid?: string
}

export interface ApiOrderESimProfile {
    actionType: 'ORDER_NEW_ESIM_PROFILE'
    eid?: string
}

export interface ApiAssignedProductToAdd {
    productId: number
    parentInventoryProductId?: number
    parameters: ApiProductParameter[]
}
export interface ApiAssignedProductToRemove {
    inventoryProductId: string
}

export interface ApiAssignedProductToChange {
    inventoryProductId: string
    parameters: ApiProductParameter[]
}

export interface ApiProductParameter {
    parameterId: number
    value: string
}